put icon markers in local folder and setup configuration. Reorganized files.
