+++ src/assets/assetConfig.js
@@ -1,0 +1,6 @@
+export default {
+  images: {
+    redMarkerIcon: require("./redMarkerIcon.png"),
+    blueMarkerICon: require("./blueMarkerIcon.png")
+  }
+};
