+++ src/screens/main.jsx
@@ -1,6 +1,6 @@
 import React, { Component } from "react";
 // import "./main.css";
-import MapComponent from "../components/map";
+import MapComponent from "../components/Map.js";
 import "../styling/styles.css";
 
 export default class main extends Component {
